modify files

+++ libs/analyse_app_time.py
@@ -12,9 +12,9 @@ summary_log_base = sys.argv[3]
 
 isExists = os.path.exists(summary_log_base)
 
-time_dic = {'SVDPlusPlus':0.00, 'PageRank':0.00, 'ShortestPaths':0.00, 'PCA':0.00, 'SVM':0.00,
-            'Terasort':0.00, 'ConnectedComponent':0.00, 'DecisionTree':0.00,
-            'TriangleCount':0.00, 'MF':0.00, 'KMeans':0.00, 'LogisticRegression':0.00}
+time_dic = {'SVDPlusPlus':0, 'PageRank':0, 'ShortestPaths':0, 'PCA':0 ,'SVM':0,
+            'Terasort':0, 'ConnectedComponent':0, 'DecisionTree':0,
+            'TriangleCount':0, 'MF':0, 'KMeans':0, 'LogisticRegression':0}
 if not isExists:
     os.mkdir(summary_log_base)
 
@@ -25,7 +25,7 @@ for filename in os.listdir(res_log_base + "/" + exp_type):
     index = 0
     while index < len(a):
         app_name = a[index][0]
-        sec = float(a[index][1].replace('\n', ''))
+        sec = int(a[index][1].replace('\n', ''))
         time_dic[app_name] += sec
         index += 1
     handeler.close()
@@ -36,7 +36,7 @@ for index in time_dic:
 summary_log_handeler = open(summary_log_path, 'w')
 summary_log_handeler.write("APP,Time\n")
 for index in time_dic:
-    summary_log_handeler.write("%s,%0.2f\n"%(index, time_dic[index]))
+    summary_log_handeler.write("%s,%d\n"%(index, time_dic[index]))
 print ("analyse finish")
 summary_log_handeler.close()
